made the caetgories drop down more dynamic
