+++ client/src/App.js
@@ -13,6 +13,7 @@ import Register from "../src/pages/Register/Register";
 import MyProduct from "./pages/MyProducts/MyProducts";
 import Dashboard from "./pages/Dashboard/Dashboard";
 import Navbar from "./components/Header/Navbar";
+import AddProductForm from "./components/AddProductForm";
 
 // Construct our main GraphQL API endpoint
 const httpLink = createHttpLink({
@@ -48,7 +49,7 @@ function App() {
                         <Route path="/" element={<Home />} />
                         <Route path="/register" element={<Register />} />
                         {/* <Route path="/dashboard" element={<Dashboard />} /> */}
-                        <Route path="/myproducts" element={<MyProduct />} />
+                        <Route path="/myproducts" element={<AddProductForm />} />
                     </Routes>
                 </div>
             </Router>
